requirements.txt
absl-py==1.4.0
aiobotocore==2.12.3
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
asn1crypto==1.5.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.1
botocore==1.34.69
cachetools==5.3.3
catboost==1.2.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
entrypoints==0.4
Faker==25.0.1
favicon==0.7.0
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.12.2
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.0
graphviz==0.20.3
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kiwisolver==1.4.5
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
plotly==5.22.0
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.8.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
pyzbar==0.1.9
referencing==0.35.1
replicate==0.25.2
requests==2.31.0
requests-toolbelt==0.10.1
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3fs==2023.12.2
scikit-learn==1.3.2
scipy==1.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.10.0
snowflake-ml-python==1.5.0
snowflake-snowpark-python==1.15.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
st-annotated-text==4.0.1
streamlit==1.34.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-on-Hover-tabs==1.0.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
validators==0.28.1
watchdog==4.0.0
wrapt==1.16.0
xgboost==1.7.6
yarl==1.9.4
